+++ tasks/msteams/src/main/java/com/walmartlabs/concord/plugins/msteams/Constants.java
@@ -26,26 +26,45 @@ public class Constants {
     public static final int DEFAULT_SO_TIMEOUT = 30_000;
     public static final int DEFAULT_RETRY_COUNT = 5;
 
-    public static final String VAR_TEAMS_PARAMS = "teamsParams";
-    public static final String VARS_DEFAULT_THEME_COLOR = "11B00A";
-    public static final String VAR_TEAM_ID = "teamId";
-    public static final String VAR_TENANT_ID = "tenantId";
-    public static final String VAR_WEBHOOK_URL = "webhookUrl";
-    public static final String VAR_ROOT_WEBHOOK_URL = "rootWebhookUrl";
-    public static final String VAR_WEBHOOK_ID = "webhookId";
-    public static final String VAR_WEBHOOKTYPE_ID = "webhookTypeId";
+    public static final String MSTEAMS_PARAMS_KEY = "msteamsParams";
+    public static final String DEFAULT_THEME_COLOR = "11B00A";
+    public static final String TEAM_ID_KEY = "teamId";
+    public static final String TENANT_ID_KEY = "tenantId";
+    public static final String WEBHOOK_URL_KEY = "webhookUrl";
+    public static final String ROOT_WEBHOOK_URL_KEY = "rootWebhookUrl";
+    public static final String WEBHOOK_ID_KEY = "webhookId";
+    public static final String WEBHOOKTYPE_ID_KEY = "webhookTypeId";
+    public static final String ACTION_KEY = "action";
+    public static final String MESSAGE_TITLE_KEY = "title";
+    public static final String MESSAGE_TEXT_KEY = "text";
+    public static final String MESSAGE_THEME_COLOR_KEY = "themeColor";
+    public static final String MESSAGE_SECTIONS_KEY = "sections";
+    public static final String MESSAGE_POTENTIAL_ACTION_KEY = "potentialAction";
+    public static final String IGNORE_ERRORS_KEY = "ignoreErrors";
 
-    public static final String VAR_PROXY_ADDRESS = "proxyAddress";
-    public static final String VAR_PROXY_PORT = "proxyPort";
-    public static final String VAR_CONNECTION_TIMEOUT = "connectTimeout";
-    public static final String VAR_SO_TIMEOUT = "soTimeout";
-    public static final String VAR_RETRY_COUNT = "retryCount";
+    public static final String PROXY_ADDRESS_KEY = "proxyAddress";
+    public static final String PROXY_PORT_KEY = "proxyPort";
+    public static final String CONNECTION_TIMEOUT_KEY = "connectTimeout";
+    public static final String SO_TIMEOUT_KEY = "soTimeout";
+    public static final String RETRY_COUNT_KEY = "retryCount";
 
     public static final int TOO_MANY_REQUESTS_ERROR = 429;
     public static final int TEAMS_SUCCESS_STATUS_CODE = 200;
     public static final int DEFAULT_RETRY_AFTER = 1;
     public static final int DEFAULT_PROXY_PORT = 8080;
 
+    public static final String[] ALL_IN_PARAMS = {
+            Constants.ACTION_KEY,
+            Constants.TEAM_ID_KEY,
+            Constants.WEBHOOK_ID_KEY,
+            Constants.WEBHOOK_URL_KEY,
+            Constants.MESSAGE_TITLE_KEY,
+            Constants.MESSAGE_TEXT_KEY,
+            Constants.MESSAGE_THEME_COLOR_KEY,
+            Constants.MESSAGE_SECTIONS_KEY,
+            Constants.MESSAGE_POTENTIAL_ACTION_KEY,
+            Constants.IGNORE_ERRORS_KEY
+    };
 
     private Constants() {
     }